Bug/87238 save preservation timeout (#641)
* Removed IProjectRepository.GetAllProjectsOnlyAsync (not in use)

* Replaced all use of IProjectRepository.GetTagByTagIdAsync which needed preservition history only to use GetTagWithPreservationHistoryByTagIdAsync

* Removed IProjectRepository.GetTagByTagIdAsync, replaced with GetTagWithActionsByTagIdAsync and GetTagWithAttachmentsHistoryByTagIdAsync

* Replaced IProjectRepository.GetTagsByTagIdsAsync with GetTagsWithPreservationHistoryByTagIdsAsync and GetTagsOnlyByTagIdsAsync.
Changed ProjectRepository.DefaultQuery to just have preservation history

* Integration tests for UpdateTag

* UpdateTag_AsPreserver_ShouldChangeTag as Preserver

* Integration tests for Void and Unvoid

* as Planner

+++ src/Equinor.ProCoSys.Preservation.Infrastructure/Repositories/ProjectRepository.cs
@@ -12,16 +12,12 @@ namespace Equinor.ProCoSys.Preservation.Infrastructure.Repositories
         public ProjectRepository(PreservationContext context)
             : base(context, context.Projects,
                 context.Projects
+                    // include preservation history only in DefaultQuery. Don't add other includes as Actions/Attachments due to performance
                     .Include(p => p.Tags)
-                    .ThenInclude(t => t.Requirements)
-                    .ThenInclude(r => r.PreservationPeriods)
-                    .ThenInclude(pp => pp.FieldValues)
-                    .ThenInclude(fv => fv.FieldValueAttachment)
-                    .Include(p => p.Tags)
-                    .ThenInclude(t => t.Actions)
-                    .ThenInclude(a => a.Attachments)
-                    .Include(p => p.Tags)
-                    .ThenInclude(t => t.Attachments)
+                        .ThenInclude(t => t.Requirements)
+                        .ThenInclude(r => r.PreservationPeriods)
+                        .ThenInclude(pp => pp.FieldValues)
+                        .ThenInclude(fv => fv.FieldValueAttachment)
             )
         {
         }
@@ -32,19 +28,44 @@ namespace Equinor.ProCoSys.Preservation.Infrastructure.Repositories
         public Task<Project> GetProjectWithTagsByNameAsync(string projectName)
             => Set.Include(p => p.Tags).SingleOrDefaultAsync(p => p.Name == projectName);
 
-        public Task<Tag> GetTagByTagIdAsync(int tagId)
+        public Task<Tag> GetTagOnlyByTagIdAsync(int tagId)
+            => Set
+                .Include(p => p.Tags)
+                .SelectMany(project => project.Tags)
+                .SingleOrDefaultAsync(tag => tag.Id == tagId);
+
+        public Task<Tag> GetTagWithPreservationHistoryByTagIdAsync(int tagId)
             => DefaultQuery
                 .SelectMany(project => project.Tags)
                 .SingleOrDefaultAsync(tag => tag.Id == tagId);
 
-        public Task<List<Tag>> GetTagsByTagIdsAsync(IEnumerable<int> tagIds)
-            => DefaultQuery
+        public Task<Tag> GetTagWithActionsByTagIdAsync(int tagId)
+            => Set
+                .Include(p => p.Tags)
+                    .ThenInclude(t => t.Actions)
+                    .ThenInclude(a => a.Attachments)
+                .SelectMany(project => project.Tags)
+                .SingleOrDefaultAsync(tag => tag.Id == tagId);
+
+        public Task<Tag> GetTagWithAttachmentsByTagIdAsync(int tagId)
+            => Set
+                .Include(p => p.Tags)
+                    .ThenInclude(t => t.Attachments)
+                .SelectMany(project => project.Tags)
+                .SingleOrDefaultAsync(tag => tag.Id == tagId);
+
+        public Task<List<Tag>> GetTagsOnlyByTagIdsAsync(IEnumerable<int> tagIds)
+            => Set
+                .Include(p => p.Tags)
                 .SelectMany(project => project.Tags)
                 .Where(tag => tagIds.Contains(tag.Id))
                 .ToListAsync();
 
-        public Task<List<Project>> GetAllProjectsOnlyAsync()
-            => Set.ToListAsync();
+        public Task<List<Tag>> GetTagsWithPreservationHistoryByTagIdsAsync(IEnumerable<int> tagIds)
+            => DefaultQuery
+                .SelectMany(project => project.Tags)
+                .Where(tag => tagIds.Contains(tag.Id))
+                .ToListAsync();
 
         public Task<List<Tag>> GetStandardTagsInProjectOnlyAsync(string projectName)
             => Set.Where(project => project.Name == projectName)
@@ -54,6 +75,14 @@ namespace Equinor.ProCoSys.Preservation.Infrastructure.Repositories
 
         public void RemoveTag(Tag tag)
         {
+            if (tag.Actions.Any())
+            {
+                throw new Exception("Can't remove Tag with any actions");
+            }
+            if (tag.Attachments.Any())
+            {
+                throw new Exception("Can't remove Tag with any attachments");
+            }
             foreach (var tagRequirement in tag.Requirements)
             {
                 _context.TagRequirements.Remove(tagRequirement);
@@ -67,7 +96,7 @@ namespace Equinor.ProCoSys.Preservation.Infrastructure.Repositories
                 .Where(tag => tag.TagType == TagType.Standard && tagNos.Contains(tag.TagNo) && stepIds.Contains(tag.StepId))
                 .ToListAsync();
 
-        public Task<Project> GetProjectByTagIdAsync(int tagId)
+        public Task<Project> GetProjectAndTagWithPreservationHistoryByTagIdAsync(int tagId)
             => DefaultQuery
                 .Where(project => project.Tags.Any(tag => tag.Id == tagId))
                 .SingleOrDefaultAsync();
